Accomodate use of some internal APIs in plug-ins
Use of functions declared in hts_internal.h has crept into two
plug-ins: hfile_s3 uses isalnum_c(), etc. and hfile_libcurl uses
the JSON parsing API.  This makes these APIs more fixed than is
suggested by their being defined in the internal header files
and not the public ones.

Move the function declarations to a new file (textutils_internal.h)
which includes a note that the functions may be used in plug-ins.
This file is included from both hts_internal.h and hfile_internal.h.
Alter hfile_s3.c and hfile_libcurl.c to no longer include
hts_internal.h.

The definition of struct hts_json_token is left in hts_internal.h.
New functions to allocate this struct and access the contents are
added with declarations in textutils_internal.h.  This is to
enable changes to the structure in the future without breaking
existing plug-ins that use it.