Fix end-of-range check and setting fd->eof in cram_next_slice()
In the loop that skips over containers in range queries, check
that the range and container references are the same before
comparing the start and end positions. Also check that
range.refid is not -1 (unmapped), as in that case the start
and end positions may not be valid.

Ensure that fd->eof is set when making range requests and the
container found does not have the desired reference id.

Fixes #651