Bug fix to free and subsequent reuse of references with -T ref.fa.
When discarding a reference and subsequently requerying it again we
set the length to 0 as well as freeing and NULLing the sequence.  This
is valid for references loaded via REF_PATH (eg md5sum), but not for
ones fetched by refs_load_fai.  These set length to non-zero and this
is used as an indicator to load the sequence from the associated .fa
file.

Fixes #654