Changed the EBI ref server to https.
The EBI reference server now uses https only.  While the http one
redirects, we cannot be sure it will stay this way forever.  (There
was zero warning of the http to https conversion.)

Also added https as a recognised token in the open_path_mfile code so
that it automatically escapes the colon for us, instead of requiring
double-colon escape tricks.