Fixes crashes caused when decoding some cramtools produced CRAM files.
In cramtools it appears that the "N" cigar op doesn't update the end
position for reference extents, meaning a long N can cause the slice
header to be invalid.  If we only load the part of the reference
listed in the header, we can then read beyond the end of this buffer.