Prevent bgzf_block_write from reading past the end of the index
Fixes segfault that occurred if the data being compressed went
beyond the extents of the index in use.  This could happen if
the index was made using `bgzip -r` (reindex), as indexes made
that way do not include an entry for the EOF block.

Blocks beyond the last index entry use BGZF_MAX_BLOCK_SIZE for
the requested length.