Pre-populate static_LDFLAGS and private_LIBS with LDFLAGS
Initialize both static_LDFLAGS and private_LIBS with the user
supplied LDFLAGS in configure.  This means the LDFLAGS settings
needed to build HTSlib will be available to programs that try
to link HTSlib statically (and so will need the same LDFLAGS
settings to link the required libraries).

Samtools/BCFtools release tarballs that embed HTSlib will end
up getting all the LDFLAGS settings twice if built in this way,
but that should be harmless.