Fix a couple of minor rANS issues with handling invalid data
In rans_uncompress_O0() and rans_uncompress_O1(), change in_sz and
out_sz to unsigned integers.  Prevents attempts to allocate
negative-size memory if out_sz bit 31 gets set.

In rans_uncompress_O1(), zero column 0 of `syms` as well as row 0
to prevent uninitialized memory reads on invalid input.  This
should be sufficient protection.  The decode array is calloc'd
so any invalid rANS state will decode to 0 and the next `syms`
look-up will go to column 0 as a result.  Any look-up in the
decode array that does not return 0 will have a corresponding
entry in `syms` due to the way they are set up together.  So as
long as the unused entries in row and column 0 of `syms` have been
zeroed the rANS codec should not try to access any other
uninitialized locations.