Fix SunStudio compiler warnings
Mainly unreachable code, a couple of integer issues and anonymous
unions.  Suppressed an anonymous union warning in cram_codecs.h
as it would need a lot of changes to fix.  A bogus 'end of loop not
reached' error is suppressed in knetfile.c (the compiler does
not like macros of the form `do { return; } while (0)`).

Replaced use of `diff -q` with `cmp` in test.pl as the -q option
is not supported in Solaris-derivatives.