Modifies faidx to work on FASTQ
Adds the ability to read and index FASTQ files in a similar way
to FASTA.  Uses a combined FASTA and FASTQ reading function to
avoid duplication.

The same .fai index suffix is used for both file types, adding
an extra column for qualities in the FASTQ index.  An error will
be returned, and a message printer using hts_log_error(), if
the caller tries to get a FASTQ entry but only a FASTA index is
available.
Closes #673

Fixes samtools/samtools#804 where a missing gzi file was not built when
a fai file was present.

Changes the len error return from -2 to -3 for fai_fetch to differentiate
missing entries from out of range numbers.

Test data is from the FASTQ example files provided as supplementary
information to the open-access publication:

P.J.A. Cock, C.J. Fields, N. Goto, M.L. Heuer and
P.M. Rice (2009). The Sanger FASTQ file format for sequences with
quality scores, and the Solexa/Illumina FASTQ variants.

See tests/fastqs_README.txt for more information.
The "wrapping_original_sanger.fastq" test cases are excluded as we
require sequence and quality lines to be wrapped in the same way.