Work around colons in reference names (GRCh38 HLA).
This extends the callback function used by the hts iterator code to
the ref:start-end range parser, permitting the parser to validate
whether the whole thing matches a reference name.

This works around parsing conflicts when given range queries like
"HLA-DRB1*12:17", which is either the whole of "HLA-DRB1*12:17" (this
exists in GRCh38) or base 17 onwards of ""HLA-DRB1*12" (which does
not).

Note there are still some undecided questions here.  Do we want to
handle the special names used in the iterator at this level to?  Eg
"*" meaning unmapped data and "." meaning whole file?