ABI/API placeholder for 64-bit positions in CRAM.
This upgrades all the internal data types to 64-bit and adds I/O
functions for encoding and decoding, but doesn't change the format
itself.

There is also code with #ifdef LARGE_POS **which should not be used**
in production.  This is there simply to act as a test for the 64-bit
API in htslib iterators.

The code is mainly copied from io_lib CRAM4 experimental branch:
https://github.com/jkbonfield/io_lib/commit/1150b9c3669055723276236ffd1f720575ed36cc