Change from int64_t to hts_pos_t typedef.
Also fixed missing dependency in bcf_sr_sort.o.

CRAM is still using int64_t internally as this is referring to the
*potential* on-disk format (with -DLARGE_POS) and none of the changes
there are externally visible in the public API anyway.

Include stdio.h in hts_defs.h so the mingw __MINGW_PRINTF_FORMAT
gets defined in all the places where it's needed.