Changes faidx to use the standard region parser.
Previously it had its own custom parser which nearly, but not quite,
matched the old hts_parse_reg code in functionality.

Note this changes the fasta coordinate type from long to int, which is
a backwards step.  However it is expected this will subsequently
change to hts_pos_t in another PR.