Fix off by one bug introduced in a13ac99.
Uncompressed BAM files weren't working, which alarmingly also wasn't
visible to the test harness.

We now have checks we can read & write uncompressed BAM and CRAM.

I also noticed a redundant CRAM test of CRAM (default) and CRAM
(version 3).  The default moved to version 3 a while back, so I made
this an explicit check of 2.1 and 3.0.