Add kstring support to khash.
It can be a pain to need to create nul-terminated strings when
hashing, if we're just dealing with substrings of known length in a
longer string (eg keys in an @SQ field).

Instead of passing in a const char * pointer, we instead hash using a
kstring_t struct (not pointer, as we wish to store the length in the
hash keys array without using external memory).

Technically this wastes a bit of memory as kstring_t has m (malloced
size) as well as l (length) and s (string), however kstring is
ubiquitous in htslib so it's the logical data type to operate on.