Amendment to b39e724 reference error messages.
If we have multiple URLs in our REF_PATH and we fail to find something
in the first path, then this isn't an error - it's the way PATHs
normally work so no message should be printed.

Furthermore if we have multiple URLs and the first part fails for some
reason other than ENOENT (connection failed, protocol not supported,
etc) then it's just a warning.  (It only becomes an error when we've
finished the entire path and failed to find the reference.)