Added CRAM indexing on-the-fly during cram write.
This required some changes to the APIs for *_idx_init and *_idx_save
as with BAI/CSI we build the index in memory and write it at the end,
while CRAI it writes the index as it goes.  The index filename is now
passed in during init and cached.

Also amended bcf_hdr_write to set hfp->format correctly so we can
query this later to identify the stream as BCF instead of generic BGZF.