Added BCF/VCF indexing on-the-fly.
This is messier than it appears due to subtle differences between CSI
on BCF and CSI on VCF.

CSI operates the same on BCF and BAM in that each ##contig or @SQ header
is counted and all are indexable, irrespective of whether they have
data.  CSI on VCF however operates like tabix, so the n_ref field is
only incremented for contigs with data.  In order to work out which
one is which, it then embeds a tabix header inside the CSI meta data
field.  Thus the foo.vcf.gz.csi file is neither pure CSI nor pure TBX.

In order to handle this case when indexing on the fly, it has to spot
newly covered contigs and dynamically append to the meta data.  (It
does this only if the input format is VCF.)