Added support for SAM.gz bai and csi indices.
For now this is using sam_itr_querys2 and sam_itr_next2 new functions
as to read a SAM file we need the bam_hdr_t struct (mapping rname to
tid).  This isn't needed by BAM and the iterator bam_readrec callback
function was not designed around having the same things available as
the general purpose sam_read1 call.

Note VCF does this differently.  SAM.gz is only readable via the
synced reader and cannot be read using a normal hts iterator.  This is
perhaps an oversight.