Recognise FASTA and FASTQ files and their FAI indices
Add htsExactFormat entries for these four file types to htslib/hts.h.

Place FASTQ in the sequence_data category as it can be considered to
be a representation of unmapped SAM. FASTA on the other hand is not,
as we don't operate on it in the same way as SAM/BAM/CRAM sequence data
and in particular we wouldn't ever want hts_open() to accept it, as
then swapping filename arguments in `samtools mpileup -f foo bar` etc
would be less detectable.

Add tests verifying that FASTA and FASTQ (and SAM) can be read
line-by-line via hts_open/hts_getline/hts_close.

Fixes #719.