Bug fixes to SAM speedup.
Bug fix to kputw and negative values.  It wasn't checking the size
needed for '-' char.

Silence valgrind warnings in SAM parser as it's minimal speed impact
to do so.

Dropped the now unneeded HAVE___BUILTIN_CLZ from Makefile. (The code
still checks for it incase your compiler supports this and isn't gcc
or clang, permitting users to manually set it).