Fix possible thread pool lock-up
In tpool_worker, the gap between hts_tpool_add_result() and
grabbing the pool_m mutex again gave a short window where
the hts_tpool_process in `q` could change before it was used.
If the process in question was detached at this point, it
could result in p->q_head being set to NULL incorrectly which
stops all thread pool activity.

The fix is to use p->q_head instead.

The included program thrash_threads7 locks up within 20 seconds
in my tests on the original code, and reliably runs to
completion with the revised version.