Tidy up thread pool mutex locking
Make new functions hts_tpool_process_detach_locked() and
hts_tpool_process_shutdown_locked() which expect the pool mutex
to already be locked.  The old functions hts_tpool_process_detach()
and hts_tpool_process_shutdown() become wrappers around these
that lock the mutex, call the new function and then unlock.

Make hts_tpool_process_destroy() call the new _locked() functions.
It already has a lock on the pool mutex, so there is no need
to lock again.

In tpool_worker(), first lock the pool mutex outside the main loop.
This removes a couple of places where the lock was being released
only to be immediately grabbed again on the next iteration.  It
allows the loop condition to be changed to `while (!p->shutdown)`,
and the code to shut down the worker to be moved outside the loop.