Do not override build variables
When you set CC, CPPFLAGS and LDFLAGS with `=` the users choices will be
overridden. The user can then only override this with `make -e`.
CPPFLAGS and LDFLAGS are empty, so there is no need at all to set these.
If CC is set by the user, e.g. to another compiler, `?=` should be used
instead of `?=` to provide the default.