More sam parsing speed-ups
Use strchr for _read_token as glibc's is quicker than the uint64_t
version.

Pass to and from locations to COPY_Q_T_MINUS_N() macro, and
change the name to COPY_MINUS_N() as it no longer works
directly on the `q` and `t` pointers (it actually reads
from `p` now).  Also add a check for borrowing when subtracting
the offset.

Remove _read_token_aux().  For quality values, check for the
correct length by testing for the tab at the end of the quality
string.  This test is slightly less stringent than it used to
be, but the borrow check in COPY_MINUS_N should catch anything
that it misses. For most aux fields, it isn't necessary to know
the size of the record anyway - the exceptions are Z and H which
can use strchr() to find the tab.  The other ones are adjusted
to work when the record is terminated with a tab as well as with
a NUL.

Change calls to kputc_() to directly writing into the kstring
where it's already known that enough space is available.