Fix possible string truncation warnings
Detected on Windows after MinGW upgraded to gcc 8.2.0.

string_ndup() now uses memcpy() instead of strncpy() so it
will now copy len bytes even if the string contains a NUL.  This
could only possibly happen in one use of this function (to store
paired read names in process_one_read, if the name illegally
includes an embedded NUL), and even in that case it will not read
any uninitialised memory.