Add test/pileup option to use the mpileup interface.
test/pileup.c is rearranged a bit to avoid code duplication.

Altered the output a bit to make it more like `samtools mpileup`:

 - Prints the reference name instead of the index number
 - Prints the pileup depth column
 - Writes ref skips using '<' and '>'
 - Prints reverse complement bases in lower case

It still puts brackets around indels to make them easier to see.
Deletions and now written as -1() instead of -1(?).

Includes comment with a perl script that converts `samtools
mpileup` output to the same format.