Fix incorrect l_extranul calculation.
Commit 6eb1051 changed how the sam parser pads the query name to
be a multiple of four bytes.  An error in an if statement caused
it to use four bytes instead of zero on names that did not need
any extra padding.  This also caused l_qname to wrap around when
the name was exactly 252 characters long, leading to problems like
an out-of-bounds memory access in sam_format1().

Replace the calculation with a corrected version that also gets
rid of the if statement.