On travis, add gcc-8 tests, remove MacOS gcc and try to tidy matrix
Add travis tests that use gcc-8 on Linux.  These take about 50
seconds longer than the installed gcc due to the extra apt
install needed, so currently there are only two of them.

Remove the gcc tests on MacOS.  It wasn't really running gcc,
it was clang with a gcc front-end.

Put all the tests in the 'matrix:' section instead of having
them added implicitly.  This is so we get the right settings
for each case and can have more control over exactly which tests
get run.

The tests get queued in the same order that they appear in the
matrix section.  As Mac jobs tend to be a bit slower to start
they've been moved up a bit in the hope that they will tile
better with the Linux ones and give a faster run time overall.