Added a multi-threaded SAM writer.
This follows a similar strategy to the multi-threaded BAM writer with
a thread dedicated to pulling things off the result queue and writing
to the hfile.

ABI BREAKING:  The sam_write1 prototype has changed.  The bam_hdr_t
pointer is no longer const, so we can record the use of this header by
incrementing the reference count.  (It may be possible we can avoid
this, but some tools destroy the header prior to closing the file,
which causes crashes instead.)