Merged SAM_hdr (as bam_hrecs_t) and bam_hdr_t.
Removes bam_hrecs_t text kstring as it duplicates the existing
bam_hdr_t text field.  bam_hdr_t text is freed and set to NULL
when the new API functions are used to make an edit so that
outdated text cannot be used in error.

Remove bam_hrecs_t ref_count as bam_hdr_t now has one.

Still a lot of work to be done, but a solid first step.

Co-Authored-By: Rob Davies <rmd+git@sanger.ac.uk>