Reserve a bit in bam_pileup1_t for use by future updates
This reserves a place holder that can be used in a future mpileup
update to flag leading insertion and padding operations.  The
code to do this does not exist yet, this is just to avoid having to
break the ABI again when we get around to it.

The reserved bit is stolen from the existing aux field, which
is used by bcftools' bcf_call_gap_prep() function.  It currently
only stores 22 bits in aux, leaving a maximum of six available.
Taking one is harmless to this function, although it it possible
that third-part code that unwisely uses this field for other purposes
could be broken.  Any such code should be rewritten to use the
bam_pileup_cd struct instead, which is the supported way of storing
client data.