Add fuzz target that opens sequences files.
hts_open_fuzzer.c is a fuzz target (see https://llvm.org/docs/LibFuzzer.html).
This is an entry point passing an arbitrary data buffer into htslib. The buffer
is saved within a temporary file and opened as a sequence file.

The htslib Makefile was updated to build hts_open_fuzzer.c as a .o file.
In order to build as a standalone binary we need to link in libFuzzer,
which (I think) would require clang. Building simply as a .o file
atleast provides a sanity check that the fuzzer builds.