Improve overlap detection heuristic.
Prevent reads with long CIGARs from being incorrectly rejected
as possible overlaps.

Don't add a read to the hash table if its mate is at an earlier
position as it can't be an overlap (if it was, its mate would
be in the hash table already).

Delay calling overlap_push() until after the call to bam_endpos()
so it can access the end position of the alignment.