Add hts_resize() interface to replace hts_expand(), hts_expand0()
hts_expand() and hts_expand0() call exit(1) is the allocation
fails, which is undesirable for a library.  hts_resize() will
instead return a non-zero exit code to the caller.

A new interface is added instead of making hts_expand() return
an exit code so that it is not necessary to add error checking
to every call to hts_expand() at once.

The new macro takes pointers to the available size and the array
location, to make it more obvious that they may be updated.
(hts_expand() takes the variables themselves and dereferences
them in the macro).