Ensure CRAM block uncomp_size >= 0
The check is already done for the RAW method so add it for the
others too.  Ensuring this means it won't sign extend when
cast to a size_t.

Remove casts in cram_uncompress_block when it compares actual
to expected uncompressed size.  It's safer to do the comparison
as a size_t.