Make iterators and indexing refuse to work on plain files.
Ensure that the input is a BGZF file before using htsfp->fp.bgzf
in bcf_itr_next() and sam_itr_next().

Make indexing refuse to work on non-bgzf SAM

It might be possible to get indexes to work on plain SAM, but it
would involve lots of changes.  We also want to encourage use of
compressed files.