Fix stale pointers in cram_next_slice()
Ensure fd->ctr and c_curr are set to NULL if they point to
the just-freed container when dealing with an empty container.
This fixes a double-free in cram_close() that could be triggered
using a suitably-crafted cram file.

Bug found by libfuzzer, which managed to generate such a file.