Add CRAM indexing tests
Add M5: tags to test/index.sam so that the cram tests can find the
right references.  It's done this way rather then using the
reference fasta file so that no UR: tags are added, possibly
changing the length of the header.

Add a function to test.pl to make an MD5 based cache from
test/ce.fa.  It's done this way as checking in the reference
files would make the git repository quite a bit bigger.