Change return type of functions that can error (and nothing more).
There are numerous places where we call e.g. hts_expand, kh_init,
kh_put, malloc/calloc, and so on, but the function has void return
type.  Clearly these can fail.

What this commit specifically does NOT do is add error checking.  It
is simply a placeholder that changes the return types, and thus
technically the ABI, permitting us to do the full checks at a later
stage.  That PR will be much larger and more time consuming, hence
splitting in two.

Every function that now returns a dummy 0 (success) value now contains
a comment "FIXME: check for errs in this function" so we can grep for
these and fix them in later releases.

We also do not add explicit HTS_RESULT_USED statements, because doing
so will force us to start adding error checking in our own code;
please refer to the earlier statement about what this commit is NOT
doing.