Add hts_reglist_create method
For creating a region list from a region char array.
Use the new method in the testing code.

hts_reglist_create fills in tid rather than reg to avoid ownership
problems with the strings in reg due to the way it was historically
used by samtools.  hts_itr_regions is modified to assume that tid
is correct if reg is NULL, and hts_itr_multi_cram no longer tries
to print out reg in a warning message.