Fix index values in hash tables when removing header lines
Index values in ref_hash / rg_hash higher than the one for the
entry that is removed need to be decremented.  Not terribly
efficient, as it needs to iterate through the entire hash table.

Ensure target_name and target_len arrays are kept in synch.

Adds a bam_hdr_remove_line_key() tests.