Support for non-standard index names
This modifies the hts_open() and hts_idx_load() functions to allow
non-standard index file names.

The advantage of this approach compared to the samtools -X solution
    https://github.com/samtools/samtools/issues/485
    https://github.com/samtools/samtools/commit/651bf1469d1abd86a78b5e57c29b24a93598b0c3
is that it works transparently for all file types and all programs
without the need to add new options.

The disadvantage are possible conflicts, but the risk can be minimized by using an
unlikely delimiter. In this commmit the string ##idx## is used (see HTS_IDX_DELIM in
htslib/hts.h).

Example of usage:
    tabix file.vcf.gz##idx##crazy-index-file-name 1:10000060-10000060