Preserve line ordering when reading headers
Adds a global line order linked list for all header records.  This
allows the input header to be reconstructed exactly when passed
through the new header API.  (The one exception to this is that an
out-of-order @HD line will be lifted to the top of the header).

Headers added programatically will currently be placed immediately
after the last existing line of the same type, of at the end
of the header if no line of the given type is present.

Removes special handling for comments.

Simplifies bam_hrecs_rebuild_text() as it now just has to walk
through the global linked list.