PG tag related modifications
Remove bam_hdr_link_pg() from public API as it should not
be necessary to call it directly.

Make bam_hdr_add_pg() work when ID, PP and PN tags are passed
to it in the varargs list.

Make bam_hrecs_vadd() ignore tags passed in its varargs list
which have an empty string as value.  This allows them to be
ignored when the same tag exists in the args va_list.

Add various bam_hdr_add_pg() tests.