Fixed hts_open_format to honour compression field.
Previously if we asked for compressed text formats it refused to
compression them, instead requiring us to add "z" to the mode string.
The purpose of this function is to override the mode if we've supplied
an htsformat struct, which is what it does elsewhere with formats.

Made htsExactFormat actually hold the exact format when writing rather
than morphing "bam" to "binary" and "sam" to "text" which made
subsequent checking of opened formats rather tedious.

Also fixed a 1 byte buffer overrun in hts_open_format when given 100
character long mode string.  Further reduced this to accept 99 long in
order to be compatible with the hts_hopen implementation (it'd lose
the trailing format code before).