Fix possible stale pointers in mpileup overlap hash
Certain records (for example one with a CIGAR that contians only
I operators) could be added to the overlap hash table without
being pushed on to the pileup list.  This would result in the
hash table pointing to invalid data after the bam structure
is reused on the next call to bam_plp_push().

Fix this by delaying the call to overlap_push() until it is
known that the record will be added to the pileup.

Fixes #852 (overlap_push assertion failure)