Removed hflush from cram_flush_container2.
Fixes #883.  The change from fflush to hflush occured in 6416ebf, but
hflush isn't a direct comparison to fflush and includes fsync /
fdatasync.  The impact is a substantial reduction in speed for some
OSes. Ubuntu Bionic showed speedups from 1m50s to 0m46s for a 1.1Gb BAM
to CRAM conversion in a single thread and 1m15s to 0m16s with 4
threads.  Whereas on Ubuntu Precise I see no change in speed.

A buffer-only flush doesn't make a great deal of sense with hflush,
hence we simply remove it.

This brings the f72add4 commit into question which is multi-threaded
hflush for bgzf to avoid a long pause at the end when we close the
file.  On Bionic the flushing is now a small hit (but it may become a
win if I tested on much larger files).

I'll leave the bgzf analysis for another commit though.