Allow missing indexes to be silently ignored; API consistency
Allow missing .tbi indexes to be silently ignored and use this
capability in vcf_hdr_read() and synced_bcf_reader.  This
stops unexpected error messages about indexes from being printed
out when reading vcf files that otherwise don't need one.

To enable this, add new function hts_idx_load3() which takes
a flags field that can control both downloading and printing
of errors when the index file is missing.  The specialisations
sam_index_load3(), tbx_index_load3() and bcf_index_load3()
are updated (or added) to take the same flags and interpret them
in the same way so that the API is consistent over all formats.

Make other adjustments needed to ensure the flags get passed
around correctly.  Update functions to use the new API.  Add
documentation.

There is one difference in behaviour.  This makes bcf_index_load()
only look for a .csi suffix (previously it would try .tbi as
well).  .csi is the only index format generated for BCF files,
and .tbi can't be used on them so the only way this would have
worked would be if a .csi index had been deliberately saved
with a .tbi suffix for some reason.  (.bai might work, but it
looks like HTSlib never made them for BCF and it certainly did
not look for them).