Make header tag string pointers const
They point into a string pool, so shouldn't be edited.

Causes a small cascade of const annotations around the cram code
due to its use of MD5 tags for filenames.

Removes some legacy code that ensured a NUL-termination before
doing a hash lookup.  This is no longer needed as the strings
are always NUL-terminated.