BAM header reference dictionary consistency
BAM uses a reference dictionary to look up reference names instead
of relying on the data in the text header.  This means it's
possible to have a reference name in the dictionary without a
corresponding @SQ line.  Other inconsistencies are also possible,
like having the references in a different order between
the dictionary and the text header.

As the dictionary gives the canonical target id to name mapping,
make sam_hdr_fill_hrecs() use it to build the initial list of references
in sam_hdr_t.hrec.ref.  sam_hrecs_update_hashes() is altered to
reconcile SQ header lines found in the text header with the ones
in this initial list.  Finally new function add_stub_ref_sq_lines()
adds SQ records for any names in the target list that did not appear
in the text header.

Function sam_hrecs_update() is renamed to sam_hrecs_vupdate() as it
takes a va_list, and made static as it is only used in header.c.
A new sam_hrecs_update() is added which takes a variable number
or arguments.

Also fix a bug in sam_hdr_write() that caused it to fail when
trying to add missing @SQ lines for headers that had not been
parsed using the new API.

Includes tests with hand-crafted BAM file where text @SQ headers
have been commented out.