Precompute cigar_tab[] and make it public as bam_cigar_table[]
With all sam_hdr_t structs sharing a single CIGAR table that is
recomputed on the first sam_parse1() call for each file, there is
a small chance of rewriting it while another thread is using it.
Avoid this by writing out the shared table's contents in the
source code and making it const.

Constify sam_hdr_t::cigar_tab and mark it as deprecated.
Ensure it is always initialised to point to bam_cigar_table[].

Provide a public bam_cigar_table[] as a better alternative, made
public in part so we can add check_cigar_tab() in test/sam.c to verify
that the table has been initialised correctly.