Make sam_hdr_update_line() update hash tables, arrays
Ensure that header state is consistent after sam_hdr_update_line().
This is a bit tricky as it could be used to update a header
record name.  If this is attempted, check that the new name
is not already in use and update everything appropriately.

Attempting to rename a @PG record is refused due to the complexity
involved.

Rename useful K() macro in header.h to TYPEKEY() which is a
bit easier to search for, although regrettable gcc won't
let you use TYPEKEY("SQ") as a case label...