Restore ability to parse SAM FLAG field in hex or octal
Prior to 9f28a4f1bee95eb0842e92a8a994c34bdfb657ae, sam_parse1()
used `strtol(p, &p, 0)` to parse SAM's FLAG field, which accepted
decimal/hex/octal (and even negative values, which are not accepted
by HTSJDK or the spec, and lead to unusable unexpected FLAG values).

HTSJDK emits and accepts decimal, hexadecimal, octal, or symbolic
letter strings based on the historic standalone samtools as refined
by the Cansam library's implementation.

Because HTSJDK accepts them (and because these alternatives hope to
be added to the SAM spec in due course), restore HTSlib's ability
to parse hexadecimal and octal. Use STRTOUL64()/strtoul() so that
minus signs are not accepted, as that is unused in practice and
leads to surprising FLAG values.