Bug fix to cram index query. (PR #911)
Query on ref "*" sometimes fails if the unmapped data is in a ref -1
block (rather than ref -2 == multi-ref). This is now more commonly
triggered due to the change in switching back out of multi-ref mode.

The problem was the range start..end is 1 to 0 (a bit odd) which
causes the slices to be skipped as unmapped slices have start 0.
In addition to the old fix in commit 23fa058 (for issue
https://github.com/samtools/samtools/issues/958), we now also
reset the range query start pos.